DB test class update
