+++ QueryDB.Core.Tests/OracleTests.cs
@@ -1,0 +1,138 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Linq;
+
+namespace QueryDB.Core.Tests
+{
+    [TestClass]
+    public class OracleTests : TestBase
+    {
+
+        #region Oracle DB Tests
+
+        #region Smoke Tests
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS), TestCategory(SMOKE_TESTS)]
+        public void Test_Oracle_FetchData()
+        {
+            var selectSql = Queries.OracleQueries.Smoke.SelectSql;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql);
+            Assert.IsTrue(data.Count > 0);
+            Assert.AreEqual("oracle", data[0].ReferenceData["CURRENT_DATABASE"]);
+
+            var dbContext = new DBContext(DB.Oracle, OracleConnectionString);
+            data = dbContext.FetchData(selectSql);
+            Assert.IsTrue(data.Count > 0);
+            Assert.AreEqual("oracle", data[0].ReferenceData["CURRENT_DATABASE"]);
+        }
+
+        #endregion
+
+        #region Fetch Data Tests
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS)]
+        public void Test_Oracle_FetchData_SelectQuery()
+        {
+            var selectSql = Queries.OracleQueries.SalesDB.SelectSql;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql);
+            Assert.IsTrue(data.Count == 12);
+            var agent = data.FirstOrDefault(X => X.ReferenceData["AGENT_NAME"] == "Benjamin");
+            Assert.AreEqual("A009", agent.ReferenceData["AGENT_CODE"]);
+            Assert.AreEqual("Benjamin", agent.ReferenceData["AGENT_NAME"]);
+            Assert.AreEqual("Hampshair", agent.ReferenceData["WORKING_AREA"]);
+            Assert.AreEqual("0.11", agent.ReferenceData["COMMISSION"]);
+            Assert.AreEqual("008-22536178", agent.ReferenceData["PHONE_NO"]);
+            Assert.AreEqual("", agent.ReferenceData["COUNTRY"]);
+        }
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS)]
+        public void Test_Oracle_FetchData_SelectQuery_UpperCaseKeys()
+        {
+            var selectSql = Queries.OracleQueries.SalesDB.SelectSql;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql, upperCaseKeys: true);
+            Assert.IsTrue(data.Count == 12);
+            var agent = data.FirstOrDefault(X => X.ReferenceData["AGENT_NAME"] == "Benjamin");
+            Assert.AreEqual("A009", agent.ReferenceData["AGENT_CODE"]);
+            Assert.AreEqual("Benjamin", agent.ReferenceData["AGENT_NAME"]);
+            Assert.AreEqual("Hampshair", agent.ReferenceData["WORKING_AREA"]);
+            Assert.AreEqual("0.11", agent.ReferenceData["COMMISSION"]);
+            Assert.AreEqual("008-22536178", agent.ReferenceData["PHONE_NO"]);
+            Assert.AreEqual("", agent.ReferenceData["COUNTRY"]);
+        }
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS)]
+        public void Test_Oracle_FetchData_SelectQuery_Joins()
+        {
+            var selectSql = Queries.OracleQueries.SalesDB.SelectSql_Join;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql);
+            Assert.IsTrue(data.Count == 34);
+            var agent = data.FirstOrDefault(X => X.ReferenceData["AGENT_CODE"] == "A004" && X.ReferenceData["CUST_CODE"] == "C00006");
+            Assert.AreEqual("A004", agent.ReferenceData["AGENT_CODE"]);
+            Assert.AreEqual("Ivan", agent.ReferenceData["AGENT_NAME"]);
+            Assert.AreEqual("C00006", agent.ReferenceData["CUST_CODE"]);
+            Assert.AreEqual("Shilton", agent.ReferenceData["CUST_NAME"]);
+            Assert.AreEqual("200104", agent.ReferenceData["ORD_NUM"]);
+            Assert.AreEqual("1500", agent.ReferenceData["ORD_AMOUNT"]);
+            Assert.AreEqual("500", agent.ReferenceData["ADVANCE_AMOUNT"]);
+            Assert.AreEqual("SOD", agent.ReferenceData["ORD_DESCRIPTION"]);
+        }
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS)]
+        public void Test_Oracle_FetchData_SelectQuery_Joins_UpperCaseKeys()
+        {
+            var selectSql = Queries.OracleQueries.SalesDB.SelectSql_Join;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql, upperCaseKeys: true);
+            Assert.IsTrue(data.Count == 34);
+            var agent = data.FirstOrDefault(X => X.ReferenceData["AGENT_CODE"] == "A004" && X.ReferenceData["CUST_CODE"] == "C00006");
+            Assert.AreEqual("A004", agent.ReferenceData["AGENT_CODE"]);
+            Assert.AreEqual("Ivan", agent.ReferenceData["AGENT_NAME"]);
+            Assert.AreEqual("C00006", agent.ReferenceData["CUST_CODE"]);
+            Assert.AreEqual("Shilton", agent.ReferenceData["CUST_NAME"]);
+            Assert.AreEqual("200104", agent.ReferenceData["ORD_NUM"]);
+            Assert.AreEqual("1500", agent.ReferenceData["ORD_AMOUNT"]);
+            Assert.AreEqual("500", agent.ReferenceData["ADVANCE_AMOUNT"]);
+            Assert.AreEqual("SOD", agent.ReferenceData["ORD_DESCRIPTION"]);
+        }
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS)]
+        public void Test_Oracle_FetchData_SelectQuery_Aliases()
+        {
+            var selectSql = Queries.OracleQueries.SalesDB.SelectSql_Alias;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql);
+            Assert.IsTrue(data.Count == 34);
+            var agent = data.FirstOrDefault(X => X.ReferenceData["AGENT_CODE"] == "A004" && X.ReferenceData["CUST_CODE"] == "C00006");
+            Assert.AreEqual("A004", agent.ReferenceData["AGENT_CODE"]);
+            Assert.AreEqual("Ivan", agent.ReferenceData["AGENT"]);
+            Assert.AreEqual("Torento", agent.ReferenceData["AGENT_LOCATION"]);
+            Assert.AreEqual("C00006", agent.ReferenceData["CUST_CODE"]);
+            Assert.AreEqual("Shilton", agent.ReferenceData["CUSTOMER"]);
+            Assert.AreEqual("Torento", agent.ReferenceData["CUSTOMER_LOCATION"]);
+        }
+
+        [TestMethod]
+        [TestCategory(DB_TESTS), TestCategory(ORACLE_TESTS)]
+        public void Test_Oracle_FetchData_SelectQuery_Aliases_UpperCaseKeys()
+        {
+            var selectSql = Queries.OracleQueries.SalesDB.SelectSql_Alias;
+            var data = new DBContext(DB.Oracle, OracleConnectionString).FetchData(selectSql, upperCaseKeys: true);
+            Assert.IsTrue(data.Count == 34);
+            var agent = data.FirstOrDefault(X => X.ReferenceData["AGENT_CODE"] == "A004" && X.ReferenceData["CUST_CODE"] == "C00006");
+            Assert.AreEqual("A004", agent.ReferenceData["AGENT_CODE"]);
+            Assert.AreEqual("Ivan", agent.ReferenceData["AGENT"]);
+            Assert.AreEqual("Torento", agent.ReferenceData["AGENT_LOCATION"]);
+            Assert.AreEqual("C00006", agent.ReferenceData["CUST_CODE"]);
+            Assert.AreEqual("Shilton", agent.ReferenceData["CUSTOMER"]);
+            Assert.AreEqual("Torento", agent.ReferenceData["CUSTOMER_LOCATION"]);
+        }
+
+        #endregion
+
+        #endregion
+
+    }
+}
